Move logging to its own file
