+++ src/expressServer.ts
@@ -2,11 +2,10 @@ import * as Sentry from "@sentry/node";
 import axios from "axios";
 import cron from "cron";
 import express, {Router} from "express";
-import fs from "fs";
 import cloneDeep from "lodash/cloneDeep";
 import onFinished from "on-finished";
 import passport from "passport";
-import winston, {level} from "winston";
+import {logger, LoggingOptions, setupLogging} from "./logger";
 import {Env, setupAuth, UserModel} from "./mongooseRestFramework";
 
 const SLOW_READ_MAX = 200;
@@ -23,95 +22,6 @@ export function setupErrorLogging() {
 }
 
 export type AddRoutes = (router: Router) => void;
-
-// Setup a default console logger.
-export const logger = winston.createLogger({
-  level: "debug",
-  transports: [
-    new winston.transports.Console({
-      debugStdout: true,
-      level: "debug",
-      format: winston.format.combine(winston.format.colorize(), winston.format.simple()),
-    }),
-  ],
-});
-
-interface LoggingOptions {
-  level?: "debug" | "info" | "warn" | "error";
-  transports?: winston.transport[];
-  disableFileLogging?: boolean;
-  disableConsoleLogging?: boolean;
-  logDirectory?: string;
-}
-
-function setupLogging(options?: LoggingOptions) {
-  logger.clear();
-  if (!options?.disableConsoleLogging) {
-    logger.add(
-      new winston.transports.Console({
-        debugStdout: !options?.level || options?.level === "debug",
-        level: options?.level ?? "debug",
-        format: winston.format.combine(winston.format.colorize(), winston.format.simple()),
-      })
-    );
-  }
-  if (!options?.disableFileLogging) {
-    const logDirectory = options?.logDirectory ?? "./log";
-    if (!fs.existsSync(logDirectory)) {
-      fs.mkdirSync(logDirectory);
-    }
-
-    const FILE_LOG_DEFAULTS = {
-      colorize: false,
-      compress: true,
-      dirname: logDirectory,
-      format: winston.format.simple(),
-      // 30 days of retention
-      maxFiles: 30,
-      // 50MB max file size
-      maxSize: 1024 * 1024 * 50,
-      // Only readable by server user
-      options: {mode: 0o600},
-    };
-
-    logger.add(
-      new winston.transports.Stream({
-        ...FILE_LOG_DEFAULTS,
-        level: "error",
-        handleExceptions: true,
-        // Use stream so we can open log in append mode rather than overwriting.
-        stream: fs.createWriteStream("error.log", {flags: "a"}),
-      })
-    );
-
-    logger.add(
-      new winston.transports.Stream({
-        ...FILE_LOG_DEFAULTS,
-        level: "info",
-        // Use stream so we can open log in append mode rather than overwriting.
-        stream: fs.createWriteStream("out.log", {flags: "a"}),
-      })
-    );
-    if (!options?.level || options?.level === "debug") {
-      logger.add(
-        new winston.transports.Stream({
-          ...FILE_LOG_DEFAULTS,
-          level: "debug",
-          // Use stream so we can open log in append mode rather than overwriting.
-          stream: fs.createWriteStream("debug.log", {flags: "a"}),
-        })
-      );
-    }
-  }
-
-  if (options?.transports) {
-    for (const transport of options.transports) {
-      logger.add(transport);
-    }
-  }
-
-  logger.debug("Logger set up complete");
-}
 
 const logRequestsFinished = function(req: any, res: any, startTime: [number, number]) {
   const diff = process.hrtime(startTime);
