+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from "./mongooseRestFramework";
 export * from "./expressServer";
 export * from "./passport";
+export * from "./logger";
